restructure & documentation (new run flow)
run.default
 - move default components to run.default (from worker & runner)
 - add numpytxt postprocessor

run.runner
 - handle numpy structured scalar for spawn params

clean old remnants in run
move load_includes to util.util

add user documentation on run config options

Tested with scalar output
Fit interface not done yet
