+++ doc/config.rst
@@ -15,24 +15,8 @@ The following gives an overview of all possible parameters
     | Directory where the 'profit.yaml' file is located.
     | Default: .
 
-* template_dir:
-    | Directory where the template for simulation input is located.
-    | Default: ./template
-
-* run_dir:
-    | Directory where the single runs are generated.
-    | Default: .
-
-* runner_backend:
-    | Not implemented yet. Decide if the simulation runs on local machine or on cluster.
-    | Default: local
-
 * uq:
     | Not implemented yet.
-
-* interface:
-    | Interface for collecting the single simulation run outputs.
-    | Default: ./interface.py
 
 * files:
 
@@ -95,12 +79,117 @@ The following gives an overview of all possible parameters
         * dtype: float
 
 * run:
+    | a command (string) -> shortcut for run/command + default values
 
-    * cmd: python3 ../simulation.py
-        | Command for starting the simulation
+    * runner:
+        | the run system to use
+        | an identifier (string) -> shortcut for runner/class + default values
 
-    * ntask: 4
-        | Split simulation on several cores.
+        * class: the identifier for the run system
+            | identifier (string)
+            | default: :code:`local`
+            | choices: detailed below + user defined choices from run/include
+
+        | other options depend on the class
+
+        .. code-block:: yaml
+
+            class: local
+            parallel: 1     # maximum number of simultaneous runs (for spawn array)
+            sleep: 0        # number of seconds to sleep while polling
+
+    * interface:
+        | the runner-worker interface
+        | an identifier (string) -> shortcut for interface/class + default values
+
+        * class: the identifier for the interface
+            | identifier (string)
+            | default: :code:`memmap`
+            | choices: detailed below + user defined choices from run/include
+
+        | other options depend on the class
+
+        .. code-block:: yaml
+
+            class: memmap
+            path: interface.npy     # memory mapped interface file, relative to base directory
+            max-size: 64            # maximum number of runs, determines size of the interface file
+
+    * pre:
+        | the worker preprocessor
+        | an identifier (string) -> shortcut for pre/class + default values
+
+        * class: the identifier for the preprocessor
+            | identifier (string)
+            | default: :code:`template`
+            | choices: detailed below + user defined choices from run/include
+
+        | other options depend on the class
+
+        .. code-block:: yaml
+
+            class: template
+            path: template      # directory to copy from, relative to base directory
+
+    * post:
+        | the worker postprocessor
+        | an identifier (string) -> shortcut for post/class + default values
+
+        * class: the identifier for the postprocessor
+            | identifier (string)
+            | default: :code:`json`
+            | choices: detailed below + user defined choices from run/include
+
+        | other options depend on the class
+
+        .. code-block:: yaml
+
+            class: json
+            path: stdout    # file to read from, relative to the run directory
+
+        .. code-block:: yaml
+
+            class: numpytxt
+            path: stdout    # file to read from, relative to the run directory
+            names: "f g"    # whitespace separated list of output variables in order, default read from config/variables
+
+    * command:
+        | shell/bash command
+        | default: :code:`./simulation`
+        | the command which starts the simulation
+
+    * stdout:
+        | :code:`null` or path
+        | default: :code:`stdout`
+        | where the simulation's stdout should be redirected to (relative to run directory)
+        | :code:`null` means insertion into the worker's stdout
+
+    * stderr:
+        | :code:`null` or path
+        | default: :code:`null`
+        | where the simulation's stderr should be redirected to (relative to run directory)
+        | :code:`null` means insertion into the worker's stderr
+
+    * clean:
+        | boolean
+        | default: :code:`true`
+        | whether to clean the run directory after execution
+
+    * time:
+        | boolean
+        | default: :code:`false`
+        | whether to record the computation time and add it to the output data (using the name :code:`TIME`)
+
+    * include:
+        | path or list of paths
+        | default: empty
+        | paths to files containing custom workers (relative to the base directory or absolute)
+        | if the custom worker & runner components register themselves, their identifiers are automatically available
+
+    * custom:
+        | boolean
+        | default: :code:`false`
+        | whether to spawn the simulation directly without worker, the simulation integrates it's own interface or worker
 
 * fit:
