+++ examples/mockup/study/template/mockup.py
@@ -1,0 +1,1 @@
+../../mockup.py
