+++ profit/defaults.py
@@ -1,0 +1,7 @@
+""" main configuration file for global default values """
+
+# config/run
+
+RUN_RUNNER_CLASS = 'local'
+RUN_RUNNER_LOCAL_PARALLEL = 1
+RUN_RUNNER_LOCAL_SLEEP = 0
