+++ profit/run/runner.py
@@ -5,16 +5,12 @@ Goal: a class to manage and deploy runs
 """
 
 import os
-from os import path
-import subprocess
 import logging
 from abc import ABC, abstractmethod  # Abstract Base Class
 from collections.abc import MutableMapping
-from time import sleep
 
-from profit import pre
 from .worker import Worker, Preprocessor
-from .run import load_includes
+from profit.util import load_includes
 
 import numpy as np
 
@@ -28,7 +24,6 @@ class RunnerInterface(ABC):
         self.base_config = base_config
 
         self.dtype = []
-        self.construct_dtype()
 
     def construct_dtype(self):  # ToDo: replace with external spec? provided by Config maybe?
         self.dtype = []
@@ -37,7 +32,7 @@ class RunnerInterface(ABC):
         self.dtype += [('DONE', np.bool8), ('TIME', np.uint32)]
         for variable, spec in self.base_config['output'].items():
             if len(spec['range']) > 0:  # vector output
-                self.dtype.append((variable, (spec['dtype'], spec['range'].shape)))
+                self.dtype.append((variable, (spec['dtype'], list(spec['range'].items())[0][1].shape)))
             else:
                 self.dtype.append((variable, spec['dtype']))
 
@@ -47,12 +42,13 @@ class RunnerInterface(ABC):
         pass
 
     def prepare(self):
-        pass
+        self.construct_dtype()
 
     def clean(self):
         pass
 
     @classmethod
+    @abstractmethod
     def handle_config(cls, config, base_config):
         pass
 
@@ -69,47 +65,7 @@ class RunnerInterface(ABC):
         return cls.interfaces[item]
 
 
-@RunnerInterface.register('memmap')
-class MemmapRunnerInterface(RunnerInterface):
-    def __init__(self, *args):
-        super().__init__(*args)
-        self._memmap = None
-
-    @property
-    def data(self):
-        if self._memmap is None:
-            raise RuntimeError('no preparation done done')
-        return self._memmap
-
-    def prepare(self):
-        """ create & load interface file """
-        init_data = np.zeros(self.config['max-size'], dtype=self.dtype)
-        np.save(self.config['path'], init_data)
-
-        try:
-            self._memmap = np.load(self.config['path'], mmap_mode='r+')
-        except FileNotFoundError:
-            self.runner.logger.error(
-                f'{self.__class__.__name__} could not load {self.config["path"]} (cwd: {os.getcwd()})')
-            raise
-
-    def clean(self):
-        os.remove(self.config['path'])
-
-    @classmethod
-    def handle_config(cls, config, base_config):
-        """
-        class: memmap
-        path: interface.npy     # memory mapped interface file, relative to base directory
-        max-size: 64            # maximum number of runs, determines size of the interface file
-        """
-        if 'path' not in config:
-            config['path'] = 'interface.npy'
-        # 'path' is relative to base_dir, convert to absolute path
-        if not os.path.isabs(config['path'][0]):
-            config['path'] = os.path.abspath(os.path.join(base_config['base_dir'], config['path']))
-        if 'max-size' not in config:
-            config['max-size'] = 64
+# === Runner === #
 
 
 class Runner(ABC):
@@ -147,10 +103,23 @@ class Runner(ABC):
         """spawn a single run
 
         :param params: a mapping which defines input parameters to be set
+        :param wait: whether to wait for the run to complete
         """
-        if params is not None:
-            for key, value in params.items():
-                self.interface.data[key][self.next_run_id] = value
+        mapping = self.params2map(params)
+        for key, value in mapping.items():
+            self.interface.data[key][self.next_run_id] = value
+
+    @staticmethod
+    def params2map(params):
+        if params is None:
+            return {}
+        if isinstance(params, MutableMapping):
+            return params
+        try:
+            return {key: params[key] for key in params.dtype.names}
+        except AttributeError:
+            pass
+        raise TypeError('params are not a Mapping')
 
     @abstractmethod
     def spawn_array(self, params_array, blocking=True):
@@ -177,6 +146,15 @@ class Runner(ABC):
         ToDo: check for correct types & valid paths
         ToDo: give warnings
         """
+        if 'include' not in config:
+            config['include'] = []
+        elif isinstance(config['include'], str):
+            config['include'] = [config['include']]
+        for p, path in enumerate(config['include']):
+            if not os.path.isabs(path):
+                config['include'][p] = os.path.abspath(os.path.join(base_config['base_dir'], path))
+        load_includes(config['include'])
+
         defaults = {'runner': 'local', 'interface': 'memmap', 'custom': False}
         for key, default in defaults.items():
             if key not in config:
@@ -188,15 +166,12 @@ class Runner(ABC):
         if not isinstance(config['interface'], MutableMapping):
             config['interface'] = {'class': config['interface']}
         RunnerInterface[config['interface']['class']].handle_config(config['interface'], base_config)
-        if 'include' not in config:
-            config['include'] = []
-        elif isinstance(config['include'], str):
-            config['include'] = [config['include']]
-        load_includes(config['include'])
+
         if not config['custom']:
             Worker.handle_config(config, base_config)
 
     @classmethod
+    @abstractmethod
     def handle_subconfig(cls, config, base_config):
         pass
 
@@ -212,45 +187,3 @@ class Runner(ABC):
 
     def __class_getitem__(cls, item):
         return cls.systems[item]
-
-
-@Runner.register('local')
-class LocalRunner(Runner):
-    # ToDo: better behaviour with asyncio.create_subprocess_exec ?
-    def spawn_run(self, params=None, wait=False):
-        super().spawn_run(params, wait)
-        env = self.env.copy()
-        env['PROFIT_RUN_ID'] = str(self.next_run_id)
-        self.runs[self.next_run_id] = subprocess.Popen(['profit-worker'], env=env, cwd=self.base_config['run_dir'])
-        if wait:
-            self.runs[self.next_run_id].wait()
-            del self.runs[self.next_run_id]
-        self.next_run_id += 1
-
-    def spawn_array(self, params_array, blocking=True):
-        """ spawn an array of runs, maximum 'parallel' at the same time, blocking until all are done """
-        if not blocking:
-            raise NotImplementedError
-        for params in params_array:
-            while len(self.runs) >= self.config['runner']['parallel']:
-                sleep(self.config['runner']['sleep'])
-                self.check_runs()
-            self.spawn_run(params)
-        while len(self.runs):
-            sleep(self.config['runner']['sleep'])
-            self.check_runs()
-
-    def check_runs(self):
-        """ check the status of runs via the interface, processes are not polled """
-        for run_id, process in list(self.runs.items()):  # preserve state before deletions
-            if self.interface.data['DONE'][run_id]:
-                process.wait()  # just to make sure
-                del self.runs[run_id]
-
-    @classmethod
-    def handle_subconfig(cls, subconfig, base_config):
-        if 'parallel' not in subconfig:
-            subconfig['parallel'] = 1
-        if 'sleep' not in subconfig:
-            subconfig['sleep'] = 1
-
