+++ profit/util/util.py
@@ -130,3 +130,21 @@ def check_ndim(arr):
 class SafeDict(dict):
     def __missing__(self, key):
         return '{' + key + '}'
+
+
+def load_includes(paths):
+    """ load python modules from the specified paths """
+    import os
+    import sys
+    from importlib.util import spec_from_file_location, module_from_spec
+    import logging
+    for path in paths:
+        name = f"profit_include_{os.path.basename(path).split('.')[0]}"
+        try:
+            spec = spec_from_file_location(name, path)
+        except FileNotFoundError:
+            logging.getLogger(__name__).error(f'could not find {path} to include')
+            continue
+        module = module_from_spec(spec)
+        sys.modules[name] = module
+        spec.loader.exec_module(module)
